requirements.txt
langchain
langchain-core
langchain-text-splitters
langgraph
langchain-openai
openai
google-generativeai
requests
langchain-community
langchain-google-community
ollama
qdrant-client
langfuse
langchain-experimental
neo4j
fastapi
uvicorn
pytest
pytest-cov
pyyaml
tiktoken
